Pre-Loadable concept
I am currently building a system at work were i have a collection of WCF
services that take an extended period of time to resolve.

We use these services to initilize our data but the service itself is
obviously resolved only when when inject it.

In a different world we could add the call to the splash screen and have
it load all the data there but the design is to only request the data as
we need it, not all at once.

So the way we got round this was to call .Resolve on all our WCF
services. This just seemed a little dirty to me. I also realized that we
must not be the only ones that need to do this. so this is my attempt at
allowing for this Pre-resolve behaviour.