DefaultComponentActivator sets nonpublic props too
The reason is that it is the model's decision what properties should be injected, not activator's. By default, model excludes non-public properties. But if some customization was applied that requested injecting through non-public setter - now it requires providing a custom activator, which is not trivial to write and use (e.g. you need to select the activator for your components correctly).